Update Graph.h
I added Dijkstra's algorithm and it has a parameter of a (string s). Also I needed to find the weights of vertex and an edge so I made two unordered maps to help me do that. The thing is I don't know exactly how to get the weight in this program, so you caloud maybe help me out for this program. Thanks